Moving the scala sensor tests to the new packages
