+++ src/test/scala/com/mwz/sonar/scala/sensor/MeasuresSpec.scala
@@ -1,9 +1,33 @@
-package com.sagacify.sonar.scala
+/*
+ * Sonar Scala Plugin
+ * Copyright (C) 2018 All contributors
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ */
+package com.mwz.sonar.scala.sensor
 
+import com.mwz.sonar.scala.Scala
 import org.scalatest._
+import org.scalactic.source.Position.apply
 
+/**
+ *  Tests the Scala Sensor Metrics
+ *
+ *  @author mwz
+ */
 class MeasuresSpec extends FlatSpec with Matchers {
-
   val exampleSourceFile =
     """/*
  * Sonar Scala Plugin
@@ -24,7 +48,7 @@ class MeasuresSpec extends FlatSpec with Matchers {
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package com.sagacify.example
+package com.mwz.sonar.scala.sensor
 
 import collection.mutable.Stack
 import org.scalatest._
@@ -99,5 +123,4 @@ class ScalaSensorSpec extends FlatSpec with Matchers {
     val count = Measures.countNonCommentLines(tokens)
     assert(count == 18)
   }
-
 }
